# tests/contract/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "Derivacion": {
        "additionalProperties": false,
        "description": "Hand-off a un asesor humano con el contexto de la interacción cargado.",
        "properties": {
          "context_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Referencia del contexto guardado",
            "title": "Context Ref"
          },
          "motivo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Motivo"
          },
          "motivo_codigo": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MotivoDerivacion"
              },
              {
                "type": "null"
              }
            ]
          },
          "requerida": {
            "default": false,
            "title": "Requerida",
            "type": "boolean"
          },
          "resumen_asesor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resumen Asesor"
          },
          "score_incomprension": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score Incomprension"
          },
          "senal_disparadora": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Regla dura o score que la disparó",
            "title": "Senal Disparadora"
          }
        },
        "title": "Derivacion",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MotivoDerivacion": {
        "description": "Por qué se deriva a un asesor humano.\n\nLos cuatro primeros son reglas duras: derivan sin calcular el score.",
        "enum": [
          "PETICION_HUMANO",
          "INVARIANTE_ROTO",
          "CONCEPTO_FUERA_CATALOGO",
          "INTENCION_REGULATORIA",
          "UMBRAL_INCOMPRENSION",
          "VERIFICACION_FALLIDA",
          "NIVEL_INSUFICIENTE"
        ],
        "title": "MotivoDerivacion",
        "type": "string"
      },
...
      "PeticionDerivacion": {
        "additionalProperties": false,
        "description": "Cuerpo de ``POST /v1/derivacion``.",
        "properties": {
          "conversation_id": {
            "format": "uuid",
            "title": "Conversation Id",
            "type": "string"
          },
          "cuenta_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cuenta Id"
          },
          "motivo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Motivo"
          },
          "motivo_codigo": {
            "$ref": "#/components/schemas/MotivoDerivacion",
            "default": "PETICION_HUMANO"
          },
          "periodo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Periodo"
          },
          "utterance": {
            "default": "",
            "maxLength": 2000,
            "title": "Utterance",
            "type": "string"
          }
        },
        "required": [
          "conversation_id"
        ],
        "title": "PeticionDerivacion",
        "type": "object"
      },
...
      "RespuestaDerivacion": {
        "additionalProperties": false,
        "description": "Lo que devuelve ``POST /v1/derivacion``.",
        "properties": {
          "cola": {
            "default": "FACTURACION_104",
            "description": "Cola de atención destino",
            "title": "Cola",
            "type": "string"
          },
          "context_ref": {
            "title": "Context Ref",
            "type": "string"
          },
          "conversation_id": {
            "format": "uuid",
            "title": "Conversation Id",
            "type": "string"
          },
          "creado_en": {
            "format": "date-time",
            "title": "Creado En",
            "type": "string"
          },
          "derivacion": {
            "$ref": "#/components/schemas/Derivacion"
          },
          "factset_sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Factset Sha256"
          },
          "lineas_brief": {
            "default": 0,
            "description": "Líneas del brief (control de longitud)",
            "title": "Lineas Brief",
            "type": "integer"
          },
          "prioridad": {
            "default": "NORMAL",
            "title": "Prioridad",
            "type": "string"
          },
          "resumen_asesor": {
            "title": "Resumen Asesor",
            "type": "string"
          },
          "trace_id": {
            "title": "Trace Id",
            "type": "string"
          },
          "vigencia_min": {
            "default": 120,
            "description": "Minutos que el contexto sigue vivo",
            "title": "Vigencia Min",
            "type": "integer"
          }
        },
        "required": [
          "trace_id",
          "conversation_id",
          "derivacion",
          "context_ref",
          "resumen_asesor",
          "creado_en"
        ],
        "title": "RespuestaDerivacion",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "JWT HS256 emitido por /dev/token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "email": "concursos-cis@ulima.edu.pe",
      "name": "Equipo recibo-claro"
    },
...
    "license": {
      "name": "MIT"
    },
    "summary": "Explica la variación del recibo con cero invenciones financieras, verificadas en código y comprobables por logs.",
    "title": "recibo-claro · Explicación inteligente de recibos Movistar",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/derivacion": {
      "post": {
        "description": "Crea el contexto del hand-off y devuelve la referencia y el brief del asesor.\n\nReutiliza el FactSet del último turno de la conversación si existe; si no —por\nejemplo, el cliente pide asesor sin haber pedido explicación—, lo construye. Un\nhand-off sin hechos sería exactamente la derivación \"a ciegas\" que este proyecto\nquiere eliminar.",
        "operationId": "derivar_v1_derivacion_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PeticionDerivacion"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaDerivacion"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Deriva a un asesor humano con el contexto de la interacción",
        "tags": [
          "derivacion"
        ]
      }
    },
    "/v1/derivacion/{context_ref}": {
      "get": {
        "description": "Devuelve el contexto que dejó el hand-off.\n\nEs lo que abre el asesor del 104 al recibir la llamada: la ficha, el FactSet sellado\ny la explicación que ya se le dio al cliente, para no repetírsela.",
        "operationId": "obtener_contexto_v1_derivacion__context_ref__get",
        "parameters": [
          {
            "in": "path",
            "name": "context_ref",
            "required": true,
            "schema": {
              "title": "Context Ref",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Obtener Contexto V1 Derivacion  Context Ref  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Contexto completo de una derivación (consola del asesor)",
        "tags": [
          "derivacion"
        ]
      }
    },
...
  },
...
}